ec06: some experimenting...
- implemented an alternate part 3 algorithm (see below) to
  see if it was faster. It wasn't.

- added timing of each part to main.

My first solution uses `bisect` module on a list of mentor
positions to quickly find how many mentors are within range
of a given novice. In big-O terms, this is approximately

    O(#novices * log #mentors)

I noticed the `str.count` method that counts non overlapping
occurrences of a substring within a certain range of indexes into the
string and since this is in C I thought it might be faster than
using bisect. This results in an algorithm that is, I think,

    O(#novices * 2000 characters to scan)

A bit surprisingly to me, this is *linear* but with a large
constant factor, much larger then `log #mentors`. I still thought
it might be faster than bisect, but those functions are written
in C as well.

So hooray for O(log n) algorithms!